pyproject.toml
[project]
name = "semsearch"
version = "0.1.0"
description = "Semantic search engine for Coq theorem prover libraries"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.52.0",
    "faiss-cpu>=1.11.0",
    "fastapi>=0.115.0",
    "pydantic>=2.12.0",
    "rich>=14.0.0",
    "sentence-transformers>=4.0.0",
    "uvicorn>=0.34.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pyright>=1.1.400",
    "pytest>=8.0.0",
]

[project.scripts]
semsearch-extract = "semsearch.extract:main"
semsearch-annotate = "semsearch.annotate:main"
semsearch-embed = "semsearch.embed:main"
semsearch-serve = "semsearch.server:main"
